aster_input

Module input_handler

Structs

RegisteredInputHandlerClass

Registered input handler class that can create handlers.

Enums

ConnectError

Errors that can occur when connecting to an input device.

Traits

InputHandler

A trait that represents an individual input handler instance for a specific device.

InputHandlerClass

A trait that represents an input handler class.